pyproject.toml
[project]
name = "manage-api"
version = "0.1.0"
description = "Management API service"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "docker>=7.1.0",
    "fastapi>=0.115.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.0",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28",
    "pytest>=8.3",
]

[project.scripts]
dev = "main:dev"
hash-password = "auth:hash_password_cli"
start = "main:main"

[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["auth", "database", "main"]
